pyproject.toml
[project]
name = "11-criar-sobrescrever-tbl-nivel-brasil-dbricks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-connect==15.4.*",
    "python-dotenv>=1.0.1",
]
